+++ sdk-base/src/main/java/com/mapbox/maps/plugin/locationcomponent/LocationConsumer.kt
@@ -1,6 +1,7 @@
 package com.mapbox.maps.plugin.locationcomponent
 
 import android.animation.ValueAnimator
+import com.mapbox.common.location.LocationError
 import com.mapbox.geojson.Point
 
 /**
@@ -10,18 +11,18 @@ interface LocationConsumer {
 
   /**
    * Called whenever the location is updated.
-   * @param location - supports multiple points to create more complex animations with intermediate points.
+   * @param location supports multiple points to create more complex animations with intermediate points.
    *  Last [location] value will always be the animator target for next animation.
-   * @param options - if specified explicitly will apply current animator option to single location update animation.
+   * @param options if specified explicitly will apply current animator option to single location update animation.
    *  Otherwise default animator options will be used.
    */
   fun onLocationUpdated(vararg location: Point, options: (ValueAnimator.() -> Unit)? = null)
 
   /**
    * Called whenever the bearing is updated.
-   * @param bearing - supports multiple bearing values to create more complex animations with intermediate points.
+   * @param bearing supports multiple bearing values to create more complex animations with intermediate points.
    *  Last [bearing] value will always be the animator target for next animation.
-   * @param options - if specified explicitly will apply current animator option to single location bearing animation.
+   * @param options if specified explicitly will apply current animator option to single location bearing animation.
    *  Otherwise default animator options will be used.
    */
   fun onBearingUpdated(vararg bearing: Double, options: (ValueAnimator.() -> Unit)? = null)
@@ -37,4 +38,10 @@ interface LocationConsumer {
    * This will apply to all upcoming updates.
    */
   fun onPuckBearingAnimatorDefaultOptionsUpdated(options: ValueAnimator.() -> Unit)
+
+  /**
+   * Called whenever there is an error.
+   * @param error The actual error
+   */
+  fun onError(error: LocationError)
 }